linux: Generalize for additional processor architectures
This commit makes 2 changes that enables adding support for
new processor types:
1) the next value of the PC to use to continue after trap breakpoint
must be adjusted differently depending on the processor
architecture.  It is not always the (pc-break_size) value where pc is
from retrieved registers on all architectures.  The code was changed
to adjust the pc by a per-architecture defined value 'TRAP_OFFSET'

2) retrieving registers after receiving PTRACE_EVENT_EXIT
is not allowed on some architectures and should not be tried
on any architecture.

Signed-off-by: Dave Lerner <dave.lerner@windriver.com>
Signed-off-by: Eugene Tarassov<eugene.tarassov@windriver.com>