Leave spaces as is in TCF logger output.
Spaces were previously handled as non-printable characters which
clobbered the TCF log output when data contained spaces. Leaving
spaces unmodified greatly improves the TCF log readability.

Note that the end of string character was also handled that way
(i.e shown as a space). Changing it to an "<eos>" string so as to
differentiate it from a normal space does not improve readability
very much though so I have chosen to leave this previous behavior
unchanged.