Dwarf parser: support relocating .debug_aranges
The .debug_aranges section contains pointers to the beginning of the
dwarf tags describing the compilation unit for a given address
range, as offsets into the .debug section.  The object may also contain
rules to relocate these offsets in a .rel[a].debug_aranges section; this
fix updates the dwarf parser to apply these relocations as necessary.