Handle more ELF symbol types.
Update the symbol component to handle more ELF symbols than only global values or functions. When dealing with modules built from assembler files, the symbols' type can be NOTYPE and have a LOCAL visibility. But still, it should be possible to retrieve those symbols.
When searching a symbol by name with a null sym_ip, the search for ELF symbols in all opened symbol files is done for globals and locals. This allows to plant a breakpoint on static functions for example.

Another patch will be proposed next to handle multiple ELF symbols with the same name.