Return more than one ELF symbol.
As the code now handles more than one global ELF symbol, this commit updates the find-by-name routine to return all ELF symbols (locals and globals) with the same name.
In the case none starting address is specified, the search is performed in all opened symbol files of the context to gather all symbols with the same name. It is up to the caller to look through the returned list to get the symbol that fits its needs.