pathmap: improve rule specification flexibility wrt trailing separators
If I have a path "target:/foo/bar.so" I want to map to "/foo/bar.so", I
can use the rule:
	"target:", ""
Alternatively, I could use the rule:
	"target:", "/"
resulting in a slightly off, but usable "//foo/bar.so".

However, what I can't do is use:
	"target:/", "/"
or the less-strictly-correct:
	"target:/", ""

Since pathmappings are very often entered by users directly, they are
prone to inconsistencies, and in an effort to improve user-friendliness,
this change adds support for the last case.  In the process, it
clarifies some of the logic and provides a fast-track for cases where
the input filename matches the rule source completely.